do not use __del__ to close session
why: It is called when an object is garbage collected which happens at
some point after all references to the object have been deleted.

On error, there is propagating exception which keep a reference and make
the __del__ not called until the end of execution. In this case,
Linphone will try to register until the end of all tests, which will
trigger fail2ban